+++ src/tools/remove_surface.cc
@@ -5,11 +5,13 @@
 #include "../mesh.h"
 #include "../block.h"
 #include "tools.h"
+#include "../network.h"
 
 
 extern GLuint add_overlay_shader;
 extern GLuint remove_overlay_shader;
 extern GLuint simple_shader;
+extern ENetPeer *peer;
 
 extern ship_space *ship;
 
@@ -17,7 +19,6 @@ extern hw_mesh *surfs_hw[6];
 
 extern void
 remove_ents_from_surface(glm::ivec3 p, int face);
-
 
 struct remove_surface_tool : tool
 {
